Move relevant svelte-query selectors into `features/sources/`
